backend/pyproject.toml
[project]
name = "authentications-nosql"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bcrypt>=4.3.0",
    "certifi>=2025.8.3",
    "fastapi>=0.116.2",
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-google-genai>=2.1.12",
    "langchain-groq>=0.3.8",
    "openai-agents>=0.3.1",
    "passlib>=1.7.4",
    "pyjwt>=2.10.1",
    "pymongo>=4.15.1",
    "pypdf>=6.0.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "uvicorn>=0.35.0",
]
